+++ src/ksana_llm/models/common/common_config.h
@@ -23,7 +23,7 @@ inline void PrepareCommonModelAttributes(const nlohmann::json &config_json, Mode
   if (config_json.contains("eos_token_id") && config_json["eos_token_id"].is_array()) {
     model_config.end_ids = config_json["eos_token_id"].get<std::vector<uint32_t>>();
   } else {
-    model_config.end_ids = std::vector<uint32_t>{config_json.value("eos_token_id", 2)};
+    model_config.end_ids = std::vector<uint32_t>{static_cast<unsigned int>(config_json.value("eos_token_id", 2))};
   }
   model_config.pad_id = config_json.value("pad_token_id", 0);
   model_config.max_position_embeddings = config_json.value("max_position_embeddings", 2048);
